modifed for bigint

+++ backend/db/migrate/20250525134135_add_team_references_to_matches.rb
@@ -1,6 +1,6 @@
 class AddTeamReferencesToMatches < ActiveRecord::Migration[8.0]
   def change
-    add_column :matches, :home_team_id, :integer
-    add_column :matches, :away_team_id, :integer
+    add_column :matches, :home_team_id, :bigint
+    add_column :matches, :away_team_id, :bigint
   end
 end
